requirements.txt
numpy==2.2.1
pandas==2.2.3
PyQt6==6.8.0
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
seabreeze==2.10.1
six==1.17.0
tzdata==2025.2
